增加 Dessert

+++ src/dataset.h
@@ -23,7 +23,7 @@ void cerr_if(bool condition, Args&&... args) {
 class Dataset {
 public:
     int dim;
-    int size;
+    size_t size;
     float* data;
 
     Dataset(int dim, fs::path path) : dim(dim) {
@@ -41,7 +41,7 @@ public:
         size = fsize / ((dim + 1) * 4);
         data = new float[size * dim];
 
-        for (int i = 0; i < size; i++) {
+        for (size_t i = 0; i < size; i++) {
             in.seekg(4, std::ios::cur);
             in.read((char*)(data + i * dim), dim * 4);
         }
@@ -53,7 +53,7 @@ public:
 
 class VSSDataset : public Dataset {
 public:
-    int set_num;
+    size_t set_num;
     std::vector<const float*> set_data;
     std::vector<int> set_len;
 
@@ -73,12 +73,12 @@ public:
         in.close();
 
         set_data[0] = data;
-        for (int i = 1; i < set_num; i++) {
+        for (size_t i = 1; i < set_num; i++) {
             set_data[i] = set_data[i - 1] + set_len[i - 1] * dim;
         }
     }
 
-    std::pair<const float*, int> get_data_len(int set_id) const { return {set_data[set_id], set_len[set_id]}; }
+    std::pair<const float*, int> get_data_len(size_t set_id) const { return {set_data[set_id], set_len[set_id]}; }
 };
 
 std::vector<std::unordered_set<int>> read_groundtruth(fs::path path) {
@@ -92,10 +92,10 @@ std::vector<std::unordered_set<int>> read_groundtruth(fs::path path) {
     size_t fsize = (size_t)in.tellg();
     in.seekg(0, std::ios::beg);
 
-    int size = fsize / ((k + 1) * 4);
+    size_t size = fsize / ((k + 1) * 4);
     std::vector<std::unordered_set<int>> gts;
 
-    for (int i = 0; i < size; i++) {
+    for (size_t i = 0; i < size; i++) {
         std::vector<int> tmp(k);
         in.seekg(4, std::ios::cur);
         in.read((char*)tmp.data(), k * 4);
